Register with saga
