+++ package.json
@@ -16,9 +16,11 @@
     "redux": "^4.0.1",
     "redux-immutable-state-invariant": "^2.1.0",
     "redux-logger": "^3.0.6",
+    "redux-saga": "^0.16.0",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^0.86.0",
-    "redux-saga": "^0.16.0"
+    "superagent": "^5.1.0",
+    "toastr": "^2.1.4"
   },
   "engines": {
     "node": "v11.10.1"
