+++ src/redux/constants/actionTypes.js
@@ -1,9 +1,3 @@
-export const GET_ERRORS = 'GET_ERRORS';
-export const CLEAR_ERRORS = 'CLEAR_ERRORS';
-export const SET_CURRENT_USER = 'SET_CURRENT_USER';
-export const GET_STORES = 'GET_STORES';
-export const CREATE_STORE = 'CREATE_STORE';
-export const DELETE_STORE = 'DELETE_STORE';
 export const REGISTER_USER = 'REGISTER_USER';
 export const REGISTER_USER_SUCCESS = 'REGISTER_USER_SUCCESS';
 export const REGISTER_USER_FAILURE = 'REGISTER_USER_FAILURE';
