+++ src/redux/middleware/index.js
@@ -1,7 +1,9 @@
 import { all } from 'redux-saga/effects';
+import {watchRegisterUserAsync} from './registrationsaga';
 
 function* rootSaga() {
     yield all([
+        watchRegisterUserAsync()
     ]);
 }
 export default rootSaga;
