--- src/redux/reducers/errorReducer.js
@@ -1,14 +1,0 @@
-import { GET_ERRORS, CLEAR_ERRORS } from '../actionCreator/types';
-
-const initialState = {};
-
-export default function(state = initialState, action) {
-  switch (action.type) {
-    case GET_ERRORS:
-      return action.payload;
-    case CLEAR_ERRORS:
-      return {};
-    default:
-      return state;
-  }
-}
