+++ src/redux/reducers/registrationReducer.js
@@ -1,37 +1,45 @@
 import {
-  REGISTER_USER, REGISTER_USER_SUCCESS, REGISTER_USER_FAILURE
+  REGISTER_USER, 
+  REGISTER_USER_SUCCESS,
+  REGISTER_USER_FAILURE
 } from '../constants/actionTypes';
 
 const initialState = {
   isRegistered: false,
   isLoading: false,
-  error: {},
+  errors: {},
   user: {}
 };
 
 const registration = (state = initialState, action) => {
+
   switch (action.type) {
     case REGISTER_USER:
       return { ...state, isLoading: true, isRegistered: false };
+
     case REGISTER_USER_SUCCESS:
       return {
         ...state,
-        isRegistered: false,
+        isRegistered: true,
         isLoading: false,
-        error: {},
+        errors: {},
+        user: action.user
 
       };
+
     case REGISTER_USER_FAILURE:
       return {
         ...state,
         isRegistered: false,
         isLoading: false,
-        error: {
+        errors: {
           error: action.error
-        }
-
+        },
+        user:{}
       };
-  }
-}
-export default registration;
 
+    default:
+      return state;
+  }
+};
+export default registration;
