+++ src/services/RegistrationAPI.js
@@ -1,0 +1,13 @@
+import axios from 'axios';
+import { resolveBaseUrl } from '.';
+
+const baseUrl = resolveBaseUrl();
+
+class RegistrationAPI {
+    
+    static register(userData) {  
+        return axios.post(`${baseUrl}register/`, userData);
+    }
+}
+   
+export default RegistrationAPI;
